sdk/python-holons/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "holons"
version = "0.1.0"
description = "Organic Programming SDK for Python"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "B. ALTER"}]

dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "grpcio-reflection>=1.60.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
